refactor(flags): Move current implementation under browser directory
