+++ packages/cozy-flags/src/browser/FlagSwitcher.jsx
