+++ packages/cozy-flags/src/browser/flag.js
