+++ packages/cozy-flags/src/browser/flag.spec.js
